paids_demo/requirements.txt
torch
torchvision
torchaudio
gradio
h5py
monai
matplotlib
numpy
pillow
requests
